Format the telemetry service like forward-interop
The main script has been changed to service like in forward-interop. The
plane link class also also been changed to plane. Both are reformatted
to 2 spaces as well.

Instead of CXN_STR, the plane host and port are in separate environment
variables. The test docker-compose config has also been updated to
reflect this.

The tests has been reformated with Jest as opposed to being written in
Mocha. Coverage has also been added.

Tests no longer check JSON because it is assumed to be handle by the koa
protobuf middleware.

Misc. config files have also been copied over from the forward-interop
service.

console.error(...) has been removed in favor of the general
logger.error(...).
