+++ services/telemetry/.babelrc
@@ -1,0 +1,17 @@
+{
+  "sourceMaps": "inline",
+  "plugins": [
+    "add-module-exports",
+    "source-map-support"
+  ],
+  "presets": [
+    [
+      "env",
+      {
+        "targets": {
+          "node": "current"
+        }
+      }
+    ]
+  ]
+}
