+++ services/telemetry/.eslintrc.yml
@@ -1,0 +1,32 @@
+env:
+  es6: true
+  node: true
+plugins:
+  - jest
+extends:
+  - 'eslint:recommended'
+  - 'plugin:jest/recommended'
+parserOptions:
+  ecmaVersion: 2017
+  sourceType: module
+rules:
+  max-len:
+    - error
+    - code: 79
+    - comments: 69
+  no-trailing-spaces:
+    - error
+  eol-last:
+    - error
+  no-multiple-empty-lines:
+    - error
+    - max: 1
+  indent:
+    - error
+    - 2
+  quotes:
+    - error
+    - single
+  semi:
+    - error
+    - always
