+++ services/telemetry/.gitignore
@@ -1,8 +1,16 @@
-# Ignoring things from npm
+# Ignore things from npm.
 node_modules/
 npm-debug.log*
 package-lock.json
 
-# Ignore message and babel output
-src/messages/
+# Ignore common directories.
+src/common
+src/messages
+
+# Ignore babel and message output.
 lib/
+src/messages.js
+
+# Ignore coverage output.
+.nyc_output
+coverage
