+++ services/telemetry/Dockerfile
@@ -1,38 +1,47 @@
-ARG BASE=node:8-slim
+ARG BASE=node:8-alpine
 
-# Compile our js source
+# Compile our js source.
 FROM ${BASE} AS builder
 
 WORKDIR /builder
 
+COPY common/nodejs/package.json src/common/
 COPY telemetry/package.json .
 
 RUN npm install
 
-COPY telemetry/src/messages src/messages
+COPY common/messages/interop.proto \
+    common/messages/telemetry.proto \
+    src/messages/
 
 RUN npm run build-msg
 
+COPY common/nodejs src/common
 COPY telemetry .
 
 RUN npm run build
 
-# Making the actual image now
+# Make the actual image now.
 FROM ${BASE}
 
 WORKDIR /app
 
+ENV NODE_ENV=production
+
+COPY common/nodejs/package.json src/common/
 COPY telemetry/package.json .
+
+RUN npm install
+
+# Add in the output from the js builder above.
+COPY --from=builder /builder/lib lib
 
 COPY telemetry/bin bin
 
-RUN export NODE_ENV=production && npm install
-
-# Adding in the output from the js builder above
-COPY --from=builder /builder/lib lib
-
-ENV CXN_STR=0.0.0.0:14550
+ENV PORT=5000 \
+    PLANE_HOST='mavproxy' \
+    PLANE_PORT='14550'
 
 EXPOSE 5000
 
-CMD npm start --silent
+CMD FORCE_COLOR=1 npm start --silent
