+++ services/telemetry/test/fixtures/raw-mission-1.json
