+++ services/telemetry/test/fixtures/raw-mission-2.json
