+++ services/telemetry/test/fixtures/raw-mission-3.json
